use React to render script param
